refactor: Move findFocusedOrActiveChromeWindow to Brokk.java
